Move FileFinder to file-system subdirectory
